ra_service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
raganything>=0.1.0
lightrag-hku>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.18
httpx>=0.27.0
